+++ partial/libuptiny/jsmn.h
@@ -2,6 +2,7 @@
 #define __JSMN_H_
 
 #include <stddef.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -39,11 +40,11 @@ enum jsmnerr {
  */
 typedef struct {
 	jsmntype_t type;
-	int start;
-	int end;
-	int size;
+	int16_t start;
+	int16_t end;
+	int16_t size;
 #ifdef JSMN_PARENT_LINKS
-	int parent;
+	int16_t parent;
 #endif
 } jsmntok_t;
 
@@ -52,9 +53,9 @@ typedef struct {
  * the string being parsed now and current position in that string
  */
 typedef struct {
-	unsigned int pos; /* offset in the JSON string */
-	unsigned int toknext; /* next token to allocate */
-	int toksuper; /* superior token node, e.g parent object or array */
+	int16_t pos; /* offset in the JSON string */
+	int16_t toknext; /* next token to allocate */
+	int16_t toksuper; /* superior token node, e.g parent object or array */
 } jsmn_parser;
 
 /**
@@ -66,8 +67,8 @@ void jsmn_init(jsmn_parser *parser);
  * Run JSON parser. It parses a JSON data string into and array of tokens, each describing
  * a single JSON object.
  */
-int jsmn_parse(jsmn_parser *parser, const char *js, size_t len,
-		jsmntok_t *tokens, unsigned int num_tokens);
+int jsmn_parse(jsmn_parser *parser, const char *js, int16_t len,
+		jsmntok_t *tokens, int16_t num_tokens);
 
 #ifdef __cplusplus
 }
